Allow gtimelog to work when timelog.txt does not exist.
Only write an empty line to timelog.txt when actually adding entries,
and only if there wasn't one at the end of the file or if the current entry
is on the next (virtual) day.

Extracted virtual_midnight as a global constant instead of hardcoding it.

Removed some unused cruft.