The tasks pane can be toggled by pressing F9 and has a close button.
Getting the padding right was a bit of a nightmare; I had to reverse-engineer
(fancy word for "read the source code") the widget structure of Totem's
sidebar.

The plan is to change the tasks pane into a sidebar that can display various
kinds of things.  What kinds?  I'm sure I'll come up with something.  E.g.
a summary of work done this week for each project would be nice.