Fix UnicodeWarning when adding non-ASCII text (#20)
Fixes the warning that appears on Python 2 if you've got gtk-completion
enabled (which is the default setting) and try to add an entry with
non-ASCII text.

There was probably a logic error as well (duplicate entries might show up
in the completion popup).

The fix costs us a little extra memory usage (a Python set object), but
the strings in it should be shared with the self.history list, so it
shouldn't be much.