Exclude GUI code from coverage metrics
This way we have a number we can aim to push to 100%.

Why can't we test main.py?  Well, the goal is to have a testsuite that
runs without a working X server, and 'from gi.repository import Gtk'
kind of needs an X server.

xvfb-run could be an alternative.  Or monkey-patching gi.repository.
I'll have to think about this some more.