Load the entire history into RAM
Because it weighs peanuts and parsing is expensive.

This required some massive refactoring of TimeLog and TimeWindow
classes.  Some of it sketched in the API redesign issue (#44), some
more-or-less ad-hoc.

I should write a better commit message detailing all I've done.

I'm sure I've introduced bugs, since I was in a rush.