Everything is terrible: make it possible to add entries
Why is it terrible?  Because slow (and I don't understand why), and
because the text view scrolling is animated (and I can't turn it off)
and wrong (doesn't reach the bottom).

And don't get me started on widget properties I set in the .ui file that
are simply ignored here (but work in gtimelog.ui and main.py).