Move log view logic into a LogView class
Tie everything up with properties.

Tricky bit: 'date' and 'showing-today' properties should be updated in
sync.  If we watch for notification signals and react at once, we might
see one of the two changed and the other one not changed.  That would be
bad.  This is avoided by using queue_update() that delays the update
using GLib.idle_add().

Hacky bit: substituting my own widget instead of the one defined in
Glade, but copying all the Glade property changes into our widget.
There's a blacklist because we don't want to override custom tabs or
tag definitions in the text buffer object, and also because some
properties cause errors/warnings when copied.

Note how LogView has separate properties for hours and office_hours.
This will allow me to implement #46 someday.  The only problem is that
office_hours might need to be a float eventually...