Store the task pane width instead of position
I thought it would be a more meaningful number (and also it doesn't
change when you resize the window), but now I'm not so sure.

First of all, the width of the task pane is not width_of_window -
position_of_pane; it's some 50 pixels less (due to client side
decorations etc.).

Second, if it doesn't change when the window is resized, why am I taking
pains to update them both in the same callback?  Oh, right, maybe this
avoids performance problems with people resizing the task pane and
getting gsettings writes on every mouse movement.

Also, this is trying to solve another problem: if the user resized the
window right before closing it, the task pane position would be saved,
but the window size would still be waiting for the 500ms timeout and
would not be.  When you restarted gtimelog, you'd get a different task
pane width.  Again, there's a but.

This solution is overkill: either storing the width instead of the
position OR storing the width and position from the same callback would
be sufficient.  AND it doesn't solve everything: the last window (or
task pane) resize is lost if the user closes the window quickly enough.