Deal with broken Gio.PropertyAction on Ubuntu 14.04 LTS
The error I get is

  ValueError: Pointer arguments are restricted to integers, capsules, and None. See: https://bugzilla.gnome.org/show_bug.cgi?id=683599

I'm guessing that it's PyGI that's broken, and not GTK+ itself.  Since I
have PyGI 3.14 on my laptop (and it works fine here), but PyGI 3.12 from
Ubuntu 14.04 LTS (yes, the LTS has GTK+ 3.10 but PyGI 3.12, go figure),
let's assume we know when this was fixed.

See #52.