Work around focus stealing prevention
This workaround has bad side effects (detailed in the comments) and I'm
likely to revert it soon.

Also there's an unfortunate interaction between this and
focus-follows-mouse mode where the newly brought to top window sometimes lacks
keyboard focus, when it pops up under my mouse pointer.  And it so
happens that the hidden-behind-other-windows gtimelog window retains
keyboard focus.