Watch symlink targets for modifications
When timelog.txt is a symlink, adding a Gio file monitor on it will only
notify about changes to the symlink itself (e.g. when you make it point
to a different target), but you will not get any notifications if the
symlink's target is modified.

So, let's add a file monitor on the symlink's target explicitly.  The
new logic handles multiple levels of indirection and avoids infinite
loops, but it is not foolproof.  If you change timelog.txt to a symlink
with a different target at runtime, gtimelog will not start a new file
monitor on the new target.

Fixes #128 (with the exception mentioned above).