Use os.startfile() for "Edit log" and "Edit tasks" on MS-Windows
If run on MS-Windows the Gtk.show_uri() doesn't work correctly for
file-URIs because the opened editor seemingly can't understand the
URI-syntax. That's why - as a workaround - if running on MS-Windows
we'll use os.startfile() instead (with the filename not the URI). That
opens your text editor of choice as expected.