Implement SMTP authentication
Adds a password field to the preferences dialog.  Shuffles the fields
around so security comes before username + password.

The password will be stored in GNOME Keyring.  This makes GnomeKeyring a
required dependency (it was optional before).  I should probably talk to
the Debian maintainer about this.

Turned on the vhomogenous property in the GtkStack, not sure why it was
turned off.  (OTOH this makes the dialog have a large amount of empty
space at the bottom for no reason, which makes me unhappy.)

Fixes a bug where the security combo box would capture keyboard focus
and not let it go: https://gitlab.gnome.org/GNOME/gtk/issues/2077.

Fixes a small bug where trying to hide the tasks infobar would try to
hide the reporting infobar instead.