Change entry search to be fuzzy.
It is now only required to enter characters of the entry in the correct
order to find an entry.

Using the entries in the screenshot in the README.
'tr' would suggest the entry 'gtimelog: experimental rewrite (GH #31)'
because the 't' on index 1 is followed by an 'r' later in the entry
text.